Initial splitting of source for cdvdman and cdvdfsv
Use cd_file_t type instead of cdl_file_t, and move the extra flag entry into the RPC structure

Add ifdef guards on some include files

Move cdvdman default settings to ifdefs
