+++ modules/iopcore/cdvdman/Makefile
@@ -1,4 +1,4 @@
-IOP_OBJS = cdvdman.o streaming.o ioplib_util.o smsutils.o imports.o exports.o
+IOP_OBJS = cdvdman.o ioops.o ncmd.o scmd.o searchfile.o streaming.o ioplib_util.o smsutils.o imports.o exports.o
 USE_DEV9 ?= 0
 
 ifeq ($(USE_HDD),1)
